fix: tow num

+++ src/main/java/com/jirengu/TwoSum.java
@@ -3,10 +3,10 @@ package com.jirengu;
 public class TwoSum {
     public static void main(String[] args) {
         //请修复下面的代码使得能正确计算打印Sum: 214
-        byte num1 = 104;
-        byte num2 = 110;
-        byte sum;
-        sum = (byte)(num1 + num2);
+        int num1 = 104;
+        int num2 = 110;
+        int sum;
+        sum = num1 + num2;
         System.out.print("Sum: " + sum);
     }
 }
